// libs/compiler/lib/visitors_v2.0/visitor.d.ts
export declare abstract class Node<T extends ts.Node = ts.Node> {
    __name: string;
    kind: ts.SyntaxKind;
    flags: ts.NodeFlags;
    parent: Node;
    decorators: Decorator[];
    modifiers: Modifier[];
    node: T;
    /**
     * 其他
     */
    docs: JSDoc[];
    from: string;
    abstract visit(visitor: Visitor, context: any): any;
}
// ...
export declare class SourceFile extends Node<ts.SourceFile> {
    statements: Statement[];
    fileName: string;
    text: string;
    moduleName: string;
    languageVariant: ts.LanguageVariant;
    isDeclarationFile: boolean;
    hasNoDefaultLib: boolean;
    languageVersion: ts.ScriptTarget;
    visit(visitor: Visitor, context: any): any;
}
/**
 * 无用的
 */
export declare type Statement = ClassDeclaration | ImportDeclaration | VariableStatement | FunctionDeclaration | InterfaceDeclaration | EnumDeclaration | TypeAliasDeclaration | ReturnStatement | IfStatement | SwitchStatement | ForOfStatement | ForInStatement | FunctionBody | ExpressionStatement | BreakStatement | ThrowStatement | DebuggerStatement | ExportDeclaration | ExportAssignment | EmptyStatement | ConditionalExpression | RegularExpressionLiteral | TaggedTemplateExpression | TryStatement | ImportEqualsDeclaration | OtherStatement;
export declare class ClassDeclaration extends Node<ts.ClassDeclaration> {
    members: (ClassElement | PropertyDeclaration | ConstructorDeclaration | MethodDeclaration | SemicolonClassElement | GetAccessorDeclaration | SetAccessorDeclaration | Decorator)[];
    name: Identifier;
    typeParameters: TypeParameterDeclaration[];
    heritageClauses: HeritageClause[];
    visit(visitor: Visitor, context: any): any;
    getDecorator<T>(name: string): (visitor: Visitor) => T | undefined | null;
}
// ...
export declare class Identifier extends Node<ts.Identifier> {
    kind: ts.SyntaxKind.Identifier;
    originalKeywordKind?: ts.SyntaxKind;
    isInJSDocNamespace?: boolean;
    escapedText: ts.__String;
    text: string;
    visit(visitor: Visitor, context: any): any;
}
export declare class TypeParameterDeclaration extends Node<ts.TypeParameterDeclaration> {
    constraint: TypeNode;
    default: TypeNode;
    expression: Expression;
    name: Identifier;
    visit(visitor: Visitor, context: any): any;
}
export declare class Decorator extends Node<ts.Decorator> {
    expression: LeftHandSideExpression;
    visit(visitor: Visitor, context: any): any;
}
export declare class ShorthandPropertyAssignment extends Node<ts.ShorthandPropertyAssignment> {
    kind: ts.SyntaxKind.ShorthandPropertyAssignment;
    name: Identifier;
    questionToken?: ts.QuestionToken;
    exclamationToken?: ts.ExclamationToken;
    equalsToken?: ts.Token<ts.SyntaxKind.EqualsToken>;
    objectAssignmentInitializer?: Expression;
    visit(visitor: Visitor, context: any): any;
}
export declare class SpreadAssignment extends Node<ts.SpreadAssignment> {
    expression: Expression;
    visit(visitor: Visitor, context: any): any;
}
export declare type AccessorDeclaration = GetAccessorDeclaration | SetAccessorDeclaration;
export declare type ObjectLiteralElementLike = PropertyAssignment | ShorthandPropertyAssignment | SpreadAssignment | MethodDeclaration | AccessorDeclaration;
export declare class PropertyAssignment extends Node<ts.PropertyAssignment> {
    initializer: Expression;
    name: PropertyName;
    questionToken: ts.QuestionToken;
    visit(visitor: Visitor, context: any): any;
}
export declare class SetAccessorDeclaration extends Node<ts.SetAccessorDeclaration> {
    name: PropertyName;
    body: FunctionBody;
    asteriskToken: AsteriskToken;
    questionToken: ts.QuestionToken;
    exclamationToken: ts.ExclamationToken;
    typeParameters: TypeParameterDeclaration[];
    parameters: ParameterDeclaration[];
    type: TypeNode;
    visit(visitor: Visitor, context: any): any;
    getDecorator<T>(name: string): (visitor: Visitor) => T | undefined | null;
}
export declare class GetAccessorDeclaration extends Node<ts.GetAccessorDeclaration> {
    name: PropertyName;
    body: FunctionBody;
    asteriskToken: AsteriskToken;
    questionToken: ts.QuestionToken;
    exclamationToken: ts.ExclamationToken;
    typeParameters: TypeParameterDeclaration[];
    parameters: ParameterDeclaration[];
    type: TypeNode;
    visit(visitor: Visitor, context: any): any;
    getDecorator<T>(name: string): (visitor: Visitor) => T | undefined | null;
}
export declare class CallExpression extends Node<ts.CallExpression> {
    kind: ts.SyntaxKind.CallExpression;
    typeArguments: TypeNode[];
    expression: LeftHandSideExpression;
    arguments: Expression[];
    visit(visitor: Visitor, context: any): any;
}
export declare class ObjectLiteralExpression extends Node<ts.ObjectLiteralExpression> {
    properties: ObjectLiteralElementLike[];
    visit(visitor: Visitor, context: any): any;
}
export declare type PropertyName = Identifier | StringLiteral | NumericLiteral | ComputedPropertyName;
export declare type TypeNode = TypeReferenceNode | TupleTypeNode | UnionTypeNode | KeywordTypeNode | FunctionTypeNode | MappedTypeNode | IndexedAccessTypeNode | TypeOperatorNode | TypeLiteralNode | ArrayTypeNode | LiteralTypeNode | TypePredicateNode | ImportTypeNode | ParenthesizedTypeNode | IntersectionTypeNode | ExpressionWithTypeArguments | JSDocNullableType;
export declare type Expression = StringLiteral | NumericLiteral | ObjectLiteralExpression | ArrayLiteralExpression | Identifier | AsExpression | AwaitExpression | ArrowFunction | BooleanLiteral | CallExpression | BinaryExpression | ElementAccessExpression | PropertyAccessExpression | PrefixUnaryExpression | TemplateExpression | NewExpression | ThisExpression | NoSubstitutionTemplateLiteral | ConditionalExpression | SpreadElement | TaggedTemplateExpression | ParenthesizedExpression | RegularExpressionLiteral | OtherExpression;
export declare class OtherExpression extends Node<ts.Expression> {
    visit(visitor: Visitor, context: any): any;
}
export declare class OmittedExpression extends Node<ts.OmittedExpression> {
    kind: ts.SyntaxKind.OmittedExpression;
    visit(visitor: Visitor, context: any): any;
}
export declare class MethodDeclaration extends Node<ts.MethodDeclaration> {
    body: FunctionBody;
    name: PropertyName;
    type: TypeNode;
    parameters: ParameterDeclaration[];
    typeParameters: TypeParameterDeclaration[];
    questionToken: ts.QuestionToken;
    _functionLikeDeclarationBrand: any;
    asteriskToken: ts.AsteriskToken;
    exclamationToken: ts.ExclamationToken;
    visit(visitor: Visitor, context: any): any;
    getDecorator<T>(name: string): (visitor: Visitor) => T | undefined | null;
}
// ...
export declare class StringLiteral extends Node<ts.StringLiteral> {
    text: string;
    isUnterminated: boolean;
    hasExtendedUnicodeEscape: boolean;
    kind: ts.SyntaxKind.StringLiteral;
    visit(visitor: Visitor, context: any): any;
    create(): string;
}
export declare class NumericLiteral extends Node<ts.NumericLiteral> {
    text: string;
    isUnterminated: boolean;
    hasExtendedUnicodeEscape: boolean;
    visit(visitor: Visitor, context: any): any;
}
// ...
export declare class AsteriskToken extends Node<ts.AsteriskToken> {
    visit(visitor: Visitor, context: any): any;
}
export declare class Modifier extends Node<ts.Modifier> {
    name: 'static' | 'readonly' | 'declare' | 'protected' | 'private' | 'export' | 'default' | 'const' | 'async' | 'abstract' | 'public';
    visit(visitor: Visitor, context: any): any;
}
export declare class FunctionBody extends Node<ts.FunctionBody> {
    statements: Statement[];
    visit(visitor: Visitor, context: any): any;
}
export declare class ParameterDeclaration extends Node<ts.ParameterDeclaration> {
    dotDotDotToken: ts.DotDotDotToken;
    name: BindingName;
    type: TypeNode;
    questionToken: ts.QuestionToken;
    initializer: Expression;
    index: number;
    visit(visitor: Visitor, context: any): any;
    getDecorator<T>(name: string): (visitor: Visitor) => T | undefined | null;
}
export declare class TupleTypeNode extends Node<ts.TupleTypeNode> {
    elementTypes: TypeNode[];
    visit(visitor: Visitor, context: any): any;
}
export declare type ModuleName = Identifier | StringLiteral;
export declare type NamespaceBody = ModuleBlock | NamespaceDeclaration;
export declare type ModuleBody = NamespaceBody | JSDocNamespaceBody;
export declare type JSDocNamespaceBody = Identifier | JSDocNamespaceDeclaration;
export declare class ModuleDeclaration extends Node<ts.DeclarationStatement> {
    kind: ts.SyntaxKind.ModuleDeclaration;
    parent: ModuleBody | SourceFile;
    name: ModuleName;
    body?: ModuleBody | JSDocNamespaceDeclaration;
    visit(visitor: Visitor, context: any): any;
}
export declare class NamespaceDeclaration extends ModuleDeclaration {
    name: Identifier;
    body: NamespaceBody;
    visit(visitor: Visitor, context: any): any;
}
export declare class JSDocNamespaceDeclaration extends ModuleDeclaration {
    name: Identifier;
    body?: JSDocNamespaceBody;
    visit(visitor: Visitor, context: any): any;
}
export declare class ModuleBlock extends Node<ts.ModuleBlock> {
    kind: ts.SyntaxKind.ModuleBlock;
    statements: Statement[];
    visit(visitor: Visitor, context: any): any;
}
export declare class ImportDeclaration extends Node<ts.ImportDeclaration> {
    kind: ts.SyntaxKind.ImportDeclaration;
    moduleSpecifier: Expression;
    importClause: ImportClause;
    visit(visitor: Visitor, context: any): any;
}
export declare class VariableStatement extends Node<ts.VariableStatement> {
    kind: ts.SyntaxKind.VariableStatement;
    declarationList: VariableDeclarationList;
    visit(visitor: Visitor, context: any): any;
}
export declare class FunctionDeclaration extends Node<ts.FunctionDeclaration> {
    body: FunctionBody;
    name: Identifier;
    parameters: ParameterDeclaration[];
    typeParameters: TypeParameterDeclaration[];
    type: TypeNode;
    visit(visitor: Visitor, context: any): any;
}
export declare class InterfaceDeclaration extends Node<ts.InterfaceDeclaration> {
    members: TypeElement[];
    typeParameters: TypeParameterDeclaration[];
    name: Identifier;
    heritageClauses: HeritageClause[];
    visit(visitor: Visitor, context: any): any;
}
export declare class HeritageClause extends Node<ts.HeritageClause> {
    types: ExpressionWithTypeArguments[];
    kind: ts.SyntaxKind.HeritageClause;
    token: ts.SyntaxKind.ExtendsKeyword | ts.SyntaxKind.ImplementsKeyword;
    visit(visitor: Visitor, context: any): any;
}
export declare class EnumDeclaration extends Node<ts.EnumDeclaration> {
    members: EnumMember[];
    name: Identifier;
    visit(visitor: Visitor, context: any): any;
}
export declare class TypeAliasDeclaration extends Node<ts.TypeAliasDeclaration> {
    name: Identifier;
    type: TypeNode;
    typeParameters: TypeParameterDeclaration[];
    visit(visitor: Visitor, context: any): any;
}
// ...
export declare class ReturnStatement extends Node<ts.ReturnStatement> {
    expression: Expression;
    visit(visitor: Visitor, context: any): any;
}
export declare class AsExpression extends Node<ts.AsExpression> {
    type: TypeNode;
    expression: Expression;
    visit(visitor: Visitor, context: any): any;
}
export declare class VariableDeclarationList extends Node<ts.VariableDeclarationList> {
    declarations: VariableDeclaration[];
    visit(visitor: Visitor, context: any): any;
}
// ...
export declare class IfStatement extends Node<ts.IfStatement> {
    expression: Expression;
    thenStatement: Statement;
    elseStatement: Statement;
    visit(visitor: Visitor, context: any): any;
}
export declare class ForOfStatement extends Node<ts.ForOfStatement> {
    awaitModifier: AwaitKeywordToken;
    initializer: Expression | VariableDeclarationList;
    expression: Expression;
    visit(visitor: Visitor, context: any): any;
}
// ...
export declare class SwitchStatement extends Node<ts.SwitchStatement> {
    kind: ts.SyntaxKind.SwitchStatement;
    expression: Expression;
    caseBlock: CaseBlock;
    possiblyExhaustive: boolean;
    visit(visitor: Visitor, context: any): any;
}
// ...
export declare class ImportClause extends Node<ts.ImportClause> {
    namedBindings: NamedImportBindings;
    name: Identifier;
    visit(visitor: Visitor, context: any): any;
}
export declare type NamedImportBindings = NamespaceImport | NamedImports;
export declare class NamespaceImport extends Node<ts.NamespaceImport> {
    name: Identifier;
    visit(visitor: Visitor, context: any): any;
}
export declare class NamedImports extends Node<ts.NamedImports> {
    elements: ImportSpecifier[];
    visit(visitor: Visitor, context: any): any;
}
export declare class ImportSpecifier extends Node<ts.ImportSpecifier> {
    propertyName: Identifier;
    name: Identifier;
    visit(visitor: Visitor, context: any): any;
}
// ...
export declare class ExportDeclaration extends Node<ts.ExportDeclaration> {
    exportClause: NamedExports;
    moduleSpecifier: Expression;
    name: Identifier | NumericLiteral | StringLiteral;
    visit(visitor: Visitor, context: any): any;
}
export declare class Symbol {
    visit(visitor: Visitor, context: any): any;
}
export declare class NamedExports extends Node<ts.NamedExports> {
    elements: ExportSpecifier[];
    visit(visitor: Visitor, context: any): any;
}
export declare class ExportSpecifier extends Node<ts.ExportSpecifier> {
    propertyName: Identifier;
    name: Identifier;
    visit(visitor: Visitor, context: any): any;
}
export declare class ExportAssignment extends Node<ts.ExportAssignment> {
    kind: ts.SyntaxKind.ExportAssignment;
    isExportEquals: boolean;
    expression: Expression;
    name: Identifier | StringLiteral | NumericLiteral;
    visit(visitor: Visitor, context: any): any;
}
// ...
export declare class JSDoc extends Node<ts.JSDoc> {
    comment: string;
    tags: JSDocTag[];
    visit(visitor: Visitor, context: any): any;
}